requirements.txt
aiohttp==3.8.3  
aiosignal==1.3.1  
async-timeout==4.0.2  
attrs==22.2.0 
charset-normalizer==2.1.1  
discord.py==2.1.0  
discord-ui==5.1.6  
frozenlist==1.3.3  
idna==3.4
multidict==6.0.4
pip==22.3.1
setuptools==65.5.0
yarl==1.8.2